lambda_functions/user_queue_lambda/requirements.txt
requests
pymysql
python-dotenv
boto3
botocore
